Add Version 1 regression tests
