--- Modules/CommonTools/Tests/CommonToolsTests/CommonToolsTests.swift
@@ -1,8 +1,0 @@
-import Testing
-@testable import CommonTools
-
-@Test func example() async throws {
-    // Write your test here and use APIs like `#expect(...)` to check expected conditions.
-    // Swift Testing Documentation
-    // https://developer.apple.com/documentation/testing
-}
