+++ Modules/MetalModule/Tests/MetalModuleTests/TestPlanets.swift
@@ -1,0 +1,39 @@
+@testable import MetalModule
+
+let testPlanets: [Planet] = [
+    Planet(name: "Sun",
+           meshName: "Sun",
+           parentName: nil,
+           radius: 1,
+           distance: 0,
+           orbitSpeed: 0,
+           rotationSpeedKmSec: 0),
+    Planet(name: "Mercury",
+           meshName: "Mercury",
+           parentName: nil,
+           radius: 1,
+           distance: 0.38,
+           orbitSpeed: 0,
+           rotationSpeedKmSec: 0),
+    Planet(name: "Earth",
+           meshName: "Earth",
+           parentName: nil,
+           radius: 1,
+           distance: 1,
+           orbitSpeed: 0,
+           rotationSpeedKmSec: 0),
+    Planet(name: "Moon",
+           meshName: "Moon",
+           parentName: "Earth",
+           radius: 1,
+           distance: 0.0025,
+           orbitSpeed: 0,
+           rotationSpeedKmSec: 0),
+    Planet(name: "Mars",
+           meshName: "Mars",
+           parentName: nil,
+           radius: 1,
+           distance: 1.52,
+           orbitSpeed: 0,
+           rotationSpeedKmSec: 0)
+]
