--- OptiUniverseTests/OptiUniverseTests.swift
@@ -1,16 +1,0 @@
-//
-//  OptiUniverseTests.swift
-//  OptiUniverseTests
-//
-//  Created by max on 01.09.2025.
-//
-
-import Testing
-
-struct OptiUniverseTests {
-
-    @Test func example() async throws {
-        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
-    }
-
-}
